Add files via upload
Being a crime reporter is a very tough job, and it seems like every new crime data map introduced by a city ends up providing less information. 
The was the case in my city when they launched CityProtect. Sure, it looks nice with a clear map of all the incidents in the area. But it's a terrible, terrible system. First off, it forces the user to zoom in for about 1/2 mile before any actual crime reports populate the map. No full list is provided, so the only way to use this map is to manually go mile by mile to see what crimes are there. 
It's god awful, and I learned to hate it with a passion. A cynical person would say that cities would only use such an unfriendly system because they want to make it appear like there's less crime. I'm a cynical person. 
For this script, I used Selenium to simulate that manual activity of going through the map and collecting any crime incidents that are there. It saved a whole lot of time.